+++ CHANGELOG.md
@@ -1,7 +1,35 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [0.4.0] - 2018-08-29
+### Added
+- Added the following options
+  - `dataFile` as a constructor option to specify a custom name for the save file
+  - `legacy` as a `getItem()` option to match the legacy version of an item
+  - `variant` as a `getItem()` option to match a certain variant of an item
+
+- Added the `getLeague()` method which returns the league that was set to be used by default
+- Added `dataFile` as a constructor option to specify a custom name for the save file
+
+### Changed
+- Major code refactor, due to that bugs might occur
+- Now using jsdoc2md for documentation
+- Methods that previously used callbacks now use promises
+- The `get()` method has been renamed to `getData()`
+- The `hasDataForLeague()` method has been renamed to `hasData()`. When no league is specified, it'll return true if any data is present
+
+### Removed
+- Removed the following options
+  - `loadOnStart` option from the constructor options
+  - `save` option from the `update()` options
+
+- Removed the following methods as they didn't fit into this package. I'll release a new package that focuses on the league API soon.
+  - `getLeagues`
+  - `saveLeagues`
+  - `updateLeagues`
 
 ## [0.3.3] - 2018-08-27
 ### Changed
@@ -44,7 +72,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.2.0] - 2018-08-26
 ### Changed
 - The `getItem()` method no longer has a callback and simply returns the item data or `null` in case of invalid parameters or missing data. *I have no idea why this method had a callback in the first place, it made everything more complicated than it should be. Sometimes you get really carried away, huh?*
-- This project no longer adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html), because at this rate this project would be at version 42.0.0 in three days. I'd still say this is a minor change ;)
 
 ## [0.1.1] - 2018-08-26
 ### Fixed
@@ -54,6 +81,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - Initial release
 
+[0.4.0]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.4.0
 [0.3.3]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.3.3
 [0.3.2]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.3.2
 [0.3.1]: https://www.npmjs.com/package/poe-ninja-api-manager/v/0.3.1